+++ R/ggplot_boxplot.R
@@ -118,6 +118,8 @@ theme_boxplot <-
 #' @param stat String of "boxplot" or "identity". Defaults to "boxplot". If identity is selected, data provided must be grouped by the x_var with ymin, lower, middle, upper, ymax variables. Note "identity" does not provide outliers.
 #' @param pal Character vector of hex codes. Defaults to viridis. Use the pals package to find a suitable palette.
 #' @param width Width of the box. Defaults to 0.5.
+#' @param alpha The alpha of the fill. Defaults to 0.1. 
+#' @param line_size The size of the outlines of boxplots. Defaults to 0.5.
 #' @param title Title string. Defaults to "[Title]".
 #' @param title_wrap Number of characters to wrap the title to. Defaults to 70. Not applicable where isMobile equals TRUE.
 #' @param subtitle Subtitle string. Defaults to "[Subtitle]".
@@ -171,6 +173,8 @@ ggplot_boxplot <- function(data,
                        stat = "boxplot",
                        pal = NULL,
                        width = 0.5,
+                       alpha = 0.1,
+                       line_size = 0.5,
                        title = "[Title]",
                        title_wrap = 70,
                        subtitle = NULL,
@@ -228,7 +232,7 @@ ggplot_boxplot <- function(data,
     else if (isMobile == TRUE) font_size_body <- 14
   }
   
-  if (is.null(pal)) pal <- viridis::viridis(4)[2]
+  if (is.null(pal)) pal <- pal_default(1)
   else pal <- pal[1]
     
   plot <- ggplot(data) +
@@ -245,9 +249,13 @@ ggplot_boxplot <- function(data,
         geom_boxplot(
           aes(x = !!x_var, y = !!y_var),
           stat = stat,
+          col = pal, 
           fill = pal,
           width = width,
-          alpha = 0.9
+          size = line_size, 
+          alpha = alpha,
+          outlier.alpha = 1, 
+          outlier.size = 1
         )
     }
     else if(!rlang::quo_is_null(group_var)) {
@@ -255,9 +263,13 @@ ggplot_boxplot <- function(data,
         geom_boxplot(
           aes(x = !!x_var, y = !!y_var, group = !!group_var),
           stat = stat,
+          col = pal, 
           fill = pal,
           width = width,
-          alpha = 0.9
+          size = line_size, 
+          alpha = alpha,
+          outlier.alpha = 1, 
+          outlier.size = 1
         )
     }
   }
@@ -273,9 +285,13 @@ ggplot_boxplot <- function(data,
           ymax = .data$ymax
         ),
         stat = stat,
+        col = pal, 
         fill = pal,
         width = width,
-        alpha = 0.9
+        size = line_size, 
+        alpha = alpha,
+        outlier.alpha = 1, 
+        outlier.size = 1
       )
   }
 
@@ -412,6 +428,8 @@ ggplot_boxplot <- function(data,
 #' @param stat String of "boxplot" or "identity". Defaults to "boxplot". If identity is selected, data provided must be grouped by the x_var and facet_var with ymin, lower, middle, upper, ymax variables. Note "identity" does not provide outliers.
 #' @param pal Character vector of hex codes. Defaults to viridis. Use the pals package to find a suitable palette.
 #' @param width Width of the box. Defaults to 0.5.
+#' @param alpha The alpha of the fill. Defaults to 0.1. 
+#' @param line_size The size of the outlines of boxplots. Defaults to 0.5.
 #' @param title Title string. Defaults to "[Title]".
 #' @param title_wrap Number of characters to wrap the title to. Defaults to 70. 
 #' @param subtitle Subtitle string. Defaults to "[Subtitle]".
@@ -458,6 +476,8 @@ ggplot_boxplot_facet <-
            stat = "boxplot",
            pal = NULL,
            width = 0.5,
+           alpha = 0.1,
+           line_size = 0.5,
            title = "[Title]",
            title_wrap = 70,
            subtitle = NULL,
@@ -514,7 +534,7 @@ ggplot_boxplot_facet <-
     if(is.null(font_size_title)) font_size_title <- 11
     if(is.null(font_size_body)) font_size_body <- 10
 
-    if (is.null(pal)) pal <- viridis::viridis(4)[2]
+    if (is.null(pal)) pal <- pal_default(1)
     else pal <- pal[1]
 
     plot <- ggplot(data) +
@@ -531,9 +551,12 @@ ggplot_boxplot_facet <-
           geom_boxplot(
             aes(x = !!x_var, y = !!y_var),
             stat = stat,
+            col = pal, 
             fill = pal,
             width = width,
-            alpha = 0.9
+            size = line_size, 
+            alpha = alpha,
+            outlier.alpha = 1
           )
       }
       else if(!rlang::quo_is_null(group_var)) {
@@ -541,9 +564,13 @@ ggplot_boxplot_facet <-
           geom_boxplot(
             aes(x = !!x_var, y = !!y_var, group = !!group_var),
             stat = stat,
+            col = pal, 
             fill = pal,
             width = width,
-            alpha = 0.9
+            size = line_size, 
+            alpha = alpha,
+            outlier.alpha = 1, 
+            outlier.size = 1
           )
       }
     }
@@ -565,9 +592,13 @@ ggplot_boxplot_facet <-
             ymax = .data$ymax
           ),
           stat = stat,
+          col = pal, 
           fill = pal,
           width = width,
-          alpha = 0.9
+          size = line_size, 
+          alpha = alpha,
+          outlier.alpha = 1, 
+          outlier.size = 1
         )
     }