+++ R/ggplot_point.R
@@ -232,7 +232,7 @@ ggplot_point <- function(data,
     else if (isMobile == TRUE) font_size_body <- 14
   }
   
-  if (is.null(pal)) pal <- viridis::viridis(4)[2]
+  if (is.null(pal)) pal <- pal_default(1)
   else pal <- pal[1]
 
   plot <- ggplot(data) +
@@ -514,7 +514,7 @@ ggplot_point_col <-
         else labels <- col_labels
       }
       n_col <- length(col_cuts) - 1
-      if (is.null(pal)) pal <- viridis::viridis(n_col)
+      if (is.null(pal)) pal <- pal_default(n_col)
       else pal <- pal[1:n_col]
     }
     else if (col_method == "category") {
@@ -523,7 +523,7 @@ ggplot_point_col <-
       }
       else n_col <- length(unique(col_var_vctr))
       
-      if (is.null(pal)) pal <- viridis::viridis(n_col)
+      if (is.null(pal)) pal <- pal_default(n_col)
       else pal <- pal[1:n_col]
       
       if (is.null(col_labels)) labels <- waiver()
@@ -770,7 +770,7 @@ ggplot_point_facet <-
     if(is.null(font_size_title)) font_size_title <- 11
     if(is.null(font_size_body)) font_size_body <- 10
     
-    if (is.null(pal)) pal <- viridis::viridis(4)[2]
+    if (is.null(pal)) pal <- pal_default(1)
     else pal <- pal[1]
 
     plot <- ggplot(data) +
@@ -1070,7 +1070,7 @@ ggplot_point_col_facet <-
         else labels <- col_labels
       }
       n_col <- length(col_cuts) - 1
-      if (is.null(pal)) pal <- viridis::viridis(n_col)
+      if (is.null(pal)) pal <- pal_default(n_col)
       else pal <- pal[1:n_col]
     } 
     else if (col_method == "category") {
@@ -1079,7 +1079,7 @@ ggplot_point_col_facet <-
       }
       else n_col <- length(unique(col_var_vctr))
       
-      if (is.null(pal)) pal <- viridis::viridis(n_col)
+      if (is.null(pal)) pal <- pal_default(n_col)
       else pal <- pal[1:n_col]
       
       if (!is.null(col_labels)) labels <- col_labels